requirements.txt
requests==2.25.0
wit==6.0.0
discord.py==1.5.1
discord==1.0.1
python-dotenv==0.15.0
